run middleware on save as well, not only on autosave
