+++ scripts/apps/authoring/authoring/services/AuthoringService.ts
@@ -11,6 +11,30 @@ import {AuthoringWorkspaceService} from './AuthoringWorkspaceService';
 import {appConfig, extensions} from 'appConfig';
 import {IPublishedArticle, IArticle, IExtensionActivationResult} from 'superdesk-api';
 import {getPublishWarningConfirmModal} from '../components/publish-warning-confirm-modal';
+import {applyMiddleware as coreApplyMiddleware} from 'core/middleware';
+import {onChangeMiddleware} from '../index';
+
+export function runBeforeUpdateMiddlware(item: IArticle, orig: IArticle): Promise<IArticle> {
+    return coreApplyMiddleware(onChangeMiddleware, {item: item, original: orig}, 'item')
+        .then(() => {
+            const onUpdateFromExtensions = Object.values(extensions).map(
+                (extension) => extension.activationResult?.contributions?.authoring?.onUpdate,
+            ).filter((updateFn) => updateFn != null);
+
+            return (
+                onUpdateFromExtensions.length < 1
+                    ? Promise.resolve(item)
+                    : onUpdateFromExtensions
+                        .reduce(
+                            (current, next) => current.then(
+                                (result) => next(orig, result),
+                            ),
+                            Promise.resolve(item),
+                        )
+                        .then((nextItem) => angular.extend(item, nextItem))
+            );
+        });
+}
 
 function isReadOnly(item: IArticle) {
     return READONLY_STATES.includes(item.state);
@@ -433,57 +457,59 @@ export function AuthoringService($q, $location, api, lock, autosave, confirm, pr
      * @param {Object} origItem
      * @param {Object} item
      */
-    this.save = function saveAuthoring(origItem, item) {
-        var diff = helpers.extendItem({}, item);
-        // Finding if all the keys are dirty for real
+    this.save = function saveAuthoring(origItem, _item) {
+        return runBeforeUpdateMiddlware(_item, origItem).then((item: IArticle) => {
+            var diff = helpers.extendItem({}, item);
+            // Finding if all the keys are dirty for real
 
-        if (angular.isDefined(origItem)) {
-            angular.forEach(_.keys(diff), (key) => {
-                if (_.isEqual(diff[key], origItem[key])) {
-                    delete diff[key];
-                }
-            });
-        }
+            if (angular.isDefined(origItem)) {
+                angular.forEach(_.keys(diff), (key) => {
+                    if (_.isEqual(diff[key], origItem[key])) {
+                        delete diff[key];
+                    }
+                });
+            }
 
-        helpers.stripHtml(diff);
-        helpers.stripWhitespaces(diff);
-        helpers.cutoffPreviousRenditions(diff, origItem);
-        autosave.stop(item);
+            helpers.stripHtml(diff);
+            helpers.stripWhitespaces(diff);
+            helpers.cutoffPreviousRenditions(diff, origItem);
+            autosave.stop(item);
 
-        if (diff._etag) { // make sure we use orig item etag
-            delete diff._etag;
-        }
+            if (diff._etag) { // make sure we use orig item etag
+                delete diff._etag;
+            }
 
-        // if current document is image and it has been changed on 'media edit' we have to update the etag
-        if (origItem.type === 'picture' && item._etag != null) {
-            diff._etag = item._etag;
-        }
+            // if current document is image and it has been changed on 'media edit' we have to update the etag
+            if (origItem.type === 'picture' && item._etag != null) {
+                diff._etag = item._etag;
+            }
 
-        helpers.filterDefaultValues(diff, origItem);
+            helpers.filterDefaultValues(diff, origItem);
 
-        if (_.size(diff) > 0) {
-            return api.save('archive', origItem, diff).then((_item) => {
-                if (origItem.type === 'picture') {
-                    item._etag = _item._etag;
-                }
+            if (_.size(diff) > 0) {
+                return api.save('archive', origItem, diff).then((_item) => {
+                    if (origItem.type === 'picture') {
+                        item._etag = _item._etag;
+                    }
+                    origItem._autosave = null;
+                    origItem._autosaved = false;
+                    origItem._locked = lock.isLockedInCurrentSession(item);
+
+                    const authoringWorkspace: AuthoringWorkspaceService = $injector.get('authoringWorkspace');
+
+                    authoringWorkspace.update(origItem);
+                    return origItem;
+                });
+            }
+
+            if (origItem) {
+                // if there is nothing to save. No diff.
                 origItem._autosave = null;
                 origItem._autosaved = false;
-                origItem._locked = lock.isLockedInCurrentSession(item);
+            }
 
-                const authoringWorkspace: AuthoringWorkspaceService = $injector.get('authoringWorkspace');
-
-                authoringWorkspace.update(origItem);
-                return origItem;
-            });
-        }
-
-        if (origItem) {
-            // if there is nothing to save. No diff.
-            origItem._autosave = null;
-            origItem._autosaved = false;
-        }
-
-        return $q.when(origItem);
+            return Promise.resolve(origItem);
+        });
     };
 
     /**
