+++ scripts/apps/authoring/authoring/services/AutosaveService.ts
@@ -1,8 +1,6 @@
 import * as helpers from 'apps/authoring/authoring/helpers';
-import {applyMiddleware as coreApplyMiddleware} from 'core/middleware';
-import {onChangeMiddleware} from '../index';
-import {extensions} from 'appConfig';
 import {IArticle} from 'superdesk-api';
+import {runBeforeUpdateMiddlware} from './AuthoringService';
 
 const RESOURCE = 'archive_autosave';
 const AUTOSAVE_TIMEOUT = 3000;
@@ -55,37 +53,20 @@ export class AutosaveService {
         let id = item._id;
 
         this.timeouts[id] = $timeout(() => {
-            coreApplyMiddleware(onChangeMiddleware, {item: item, original: orig}, 'item')
-                .then(() => {
-                    const onUpdateFromExtensions = Object.values(extensions).map(
-                        (extension) => extension.activationResult?.contributions?.authoring?.onUpdate,
-                    ).filter((updateFn) => updateFn != null);
+            runBeforeUpdateMiddlware(item, orig)
+                .then((itemLatest: IArticle) => {
+                    var diff = helpers.extendItem({_id: id}, itemLatest);
 
-                    return (
-                        onUpdateFromExtensions.length < 1
-                            ? Promise.resolve(item)
-                            : onUpdateFromExtensions
-                                .reduce(
-                                    (current, next) => current.then(
-                                        (result) => next(orig, result),
-                                    ),
-                                    Promise.resolve(item),
-                                )
-                                .then((nextItem) => angular.extend(item, nextItem))
-                    ).then((itemLatest: IArticle) => {
-                        var diff = helpers.extendItem({_id: id}, itemLatest);
+                    helpers.filterDefaultValues(diff, orig);
 
-                        helpers.filterDefaultValues(diff, orig);
+                    return api.save(RESOURCE, {}, diff).then((_autosave) => {
+                        orig._autosave = _autosave;
 
-                        return api.save(RESOURCE, {}, diff).then((_autosave) => {
-                            orig._autosave = _autosave;
+                        if (typeof callback === 'function') {
+                            callback();
+                        }
 
-                            if (typeof callback === 'function') {
-                                callback();
-                            }
-
-                            return _autosave;
-                        });
+                        return _autosave;
                     });
                 });
         }, timeout, false);
